Add files via upload
task1.: If we don't use iostream header file and use cout then the preprocessor will work normally.
task2. If we use .h in a header file then it will give fatal error.
task3. If we put integer value in a character variable then data loss is observed.So we use cin.get().